requirements.txt
torch>=2.4.0
transformers>=5.5.0
accelerate>=1.1.0
bitsandbytes>=0.45.0
huggingface_hub>=0.28.0
kernels>=0.11.1
pydantic>=2.10.0
python-dotenv>=1.0.1
sentencepiece>=0.2.0
protobuf>=5.26.0
timm>=1.0.0
outlines>=0.1.1